Delete Network View fix issue #57
This fix requires a newer version of infoblox-go-client is released
with infobloxopen/infoblox-go-client#92

The fix is not a good one.  This resource stores name for the id
unlike all the other resources which use Ref off the Struct in the
go client.  Changing it now would disrupt statefiles.  Either
we live with the computed network_view_ref attribute, see if it
is possible to transition the state, or we need to deprecate this
attribute in a future API release.